% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyR4terms.R
\name{create.Myfis}
\alias{create.Myfis}
\title{Building a fuzzy inference system}
\usage{
create.Myfis()
}
\value{
a fis  object - a fuzzy  inference system.
}
\description{
This function creates a fuzzy inference system that contains three input variables and three output variables.
Each variable is expressed by 4 linguistic terms as low, medium, medium_high and high.
A trapezoidal membership function is used to match input and output to a degree of membership.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
