% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{dataFSMUMI}
\alias{dataFSMUMI}
\title{An example of a multivariate times series having three signals}
\format{An object of class \code{matrix} with 32000 rows and 3 columns.}
\usage{
dataFSMUMI
}
\description{
An example of a multivariate times series having three signals
}
\keyword{datasets}
