% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalfis.R
\name{evalfis}
\alias{evalfis}
\title{FIS evaluation}
\usage{
evalfis(input_stack, fis)
}
\arguments{
\item{input_stack}{A matrix representing the input stack, number of inputs (columns) by number of outputs (rows).}

\item{fis}{A fis must be provided.}
}
\value{
An evaluated crisp value for a given fis structure.
}
\description{
Evaluate a Fuzzy Inference System (fis). For more details, please see FuzzyR package.
}
\author{
Jon Garibaldi, Chao Chen, Tajul Razak
}
