% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbsm.R
\name{fbsm}
\alias{fbsm}
\title{Global threshold for missing data imputation}
\usage{
fbsm(Y, X)
}
\arguments{
\item{Y}{input vector}

\item{X}{input vector 
#' @examples
data(dataFSMUMI)
X <- dataFSMUMI[, 1] ; Y <- dataFSMUMI[, 2]
fbsm(Y,X)}
}
\description{
Computing the new similarity measure between two univariate signals  Y and X  based on fuzzy logic .
}
\details{
This function returns the similarity measure of two vectors \eqn{0 <= fbsm <= 3}.
fbsm is close to 3 the two vectors are more similar.
Both vectors Y and X must be of equal length, otherwise an error will be appreared.
The input vectors do not contain NA.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
\keyword{internal}
