% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding_fuzzy_based_global_threshold.R
\name{Finding_fuzzy_based_global_threshold}
\alias{Finding_fuzzy_based_global_threshold}
\title{Global threshold based on fuzzy-weighted similarity threshold}
\usage{
Finding_fuzzy_based_global_threshold(query, database, i_start, i_finish,
  step_threshold)
}
\description{
Find the global threshold using Fuzzy Logic Theory to impute large missing data in univariate time series.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
\keyword{internal}
