% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finding_fuzzy_based_similar_windows.R
\name{Finding_fuzzy_based_similar_windows}
\alias{Finding_fuzzy_based_similar_windows}
\title{Finding the similar windows using based on fuzzy-weighted similarity}
\usage{
Finding_fuzzy_based_similar_windows(query, database, i_start, i_finish,
  step_finding, threshold)
}
\description{
Find the similar windows to queries using the fuzzy-weighted similarity to impute large missing data in univariate time series.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
\keyword{internal}
