% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSMUMImputation.R
\name{FSMUMImputation}
\alias{FSMUMImputation}
\title{Imputing large gaps based on the new fuzzy-weighted similarity measure}
\usage{
FSMUMImputation(data, large_gap_threshold, step_threshold, step_finding)
}
\arguments{
\item{data}{a multivariate signals containing gaps}

\item{large_gap_threshold}{threshold used to determine a gap is large}

\item{step_threshold}{increment used for finding the threshold}

\item{step_finding}{increment used for retrieving  a similar sequences to the queries}
}
\value{
returns a completed multivariate time series
}
\description{
Fill large gaps in low or uncorrelated multivariate time series using the fuzzy-weighted similarity measure
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
