% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputing_1NA.R
\name{imp_1NA}
\alias{imp_1NA}
\title{Completing isolated missing points (1NA)}
\usage{
imp_1NA(signal, pos1)
}
\arguments{
\item{signal}{a univariate time series containing isolated missing values.}

\item{pos1}{the position of the begining of gaps of size 1,  obtained  from Indexes_size_missing() function.
This function returns a series with imputed values of all 1NAs.}
}
\description{
Complete isolated missing points  using the average of nearest neighbours.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
