% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputing_small_gaps.R
\name{imp_small_gaps_WMA}
\alias{imp_small_gaps_WMA}
\title{Completing small gaps that their sizes belong to (1,large_gap_threshold).}
\usage{
imp_small_gaps_WMA(signal, posMA, pMA, store_w, lag)
}
\arguments{
\item{posMA}{position of the begining of small gaps}

\item{pMA}{position of all corresponding values of small gaps}

\item{store_w}{list storing index and size of gaps}

\item{lag}{lag of the weighted moving average function}

\item{Data}{multivariate data}

\item{univariate}{signal}
}
\value{
returns a vector of imputed values
}
\description{
Imputing small gaps using weighted moving average.
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
\keyword{internal}
