% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indexes_size_missing.R
\name{Indexes_size_missing}
\alias{Indexes_size_missing}
\title{Indexes and sizes of gaps}
\usage{
Indexes_size_missing(data)
}
\arguments{
\item{data}{a multivariate data}
}
\value{
Return a list  per signal in which each element containing the position of the starting of a gap (1st column) and its size (2nd column).
}
\description{
Find the first positions of all gaps and their sizes respectively  within a multivariate time series.
}
\examples{
data(dataFSMUMI)
X <- dataFSMUMI
rate <- 0.1
ngaps <- 1
incompleted_signal <- Creating_gaps(X, rate,1)
id_NA <- Indexes_size_missing(incompleted_signal)
}
\author{
Thi-Thu-Hong Phan, Andre Bigand, Emilie Poisson-Caillault
}
